requirements-dev.txt
# Development dependencies for CI/CD pipeline
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5
pre-commit>=3.3.0

# Testing specific packages
factory-boy>=3.3.0
freezegun>=1.2.2
responses>=0.23.0

# Type stubs
types-requests>=2.31.0
pandas-stubs>=2.0.0

# Airflow testing
apache-airflow-providers-postgres>=5.6.0
